backend/requirements/base.txt
# Requirements base — Dependencias compartidas
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
python-decouple>=3.8
psycopg2-binary>=2.9,<3.0
django-extensions>=3.2
Pillow>=10.0
gunicorn>=21.2
argon2-cffi>=23.1
firebase-admin>=6.5,<7.0
